Update ai.py
Reformatted variables to follow snake casing. I was able to develop an algorithm to identify pieces and count the type of pieces. Still need to develop a sort of weighted utility function that uses the number of pieces a player has, how many are kings, and their position relative to the board and the distance to the row to king pawns.