Updated AI and checkers
Added section in Utility was getting stuck in an infinite loop for some reason. Commented that part out for now.
Updated search so that, if there are multiple moves with the same utility, one is selected at random rather than always selecting the last move in the dictionary.
Added sections to handle flags set in Game() function signature aside from firstmove (still unsure about the usage of that variable).
Added comments for readability.